[gui] Rename review status selector
Rename SelectReviewStatus to SelectReviewStatusDialog since it is a dialog
with an activator prop of review status selector.
