+++ web/server/vue-cli/src/components/Report/index.js
@@ -1,7 +1,7 @@
 import { ReportComments } from "./Comment";
 import Report from "./Report";
 import ReportStepMessage from "./ReportStepMessage";
-import SelectReviewStatus from "./SelectReviewStatus";
+import SelectReviewStatusDialog from "./SelectReviewStatusDialog";
 import SelectReviewStatusItem from "./SelectReviewStatusItem";
 import SelectSameReport from "./SelectSameReport";
 
@@ -9,7 +9,7 @@ export {
   Report,
   ReportComments,
   ReportStepMessage,
-  SelectReviewStatus,
+  SelectReviewStatusDialog,
   SelectReviewStatusItem,
   SelectSameReport
 };
